feature: Unit tests for constraints
