+++ chats/apps/api/v1/external/rooms/serializers.py
@@ -73,6 +73,7 @@ class RoomFlowSerializer(serializers.ModelSerializer):
                 else validated_data.pop("sector_uuid")
             )
         except AttributeError:
+            print("a")
             raise ValidationError(
                 {"detail": _("Cannot create room without queue_uuid or sector_uuid")}
             )
