+++ chats/apps/api/v1/msgs/serializers.py
@@ -47,6 +47,7 @@ class MessageMediaSerializer(serializers.ModelSerializer):
             "media_file",
             "url",
             "created_on",
+            "sender"
         ]
 
         extra_kwargs = {
