+++ chats/fixtures/fixture_app.json
@@ -2554,16 +2554,6 @@
         }
     },
     {
-        "model": "projects.flow",
-        "pk": "56336d40-ff31-49f1-ace4-edf08a0ee71e",
-        "fields": {
-            "created_on": "2022-09-02T20:39:55.075Z",
-            "modified_on": "2022-09-02T20:39:55.075Z",
-            "project_flows_uuid": "20f82849-5189-40d1-bd93-62adaaa644d2",
-            "project": "34a93b52-231e-11ed-861d-0242ac120002"
-        }
-    },
-    {
         "model": "authtoken.token",
         "pk": "1218da72b087b8be7f0e2520a515e968ab866fdd",
         "fields": {
